refatorando, ajustando obejto operações e tela de ativo

+++ app/src/main/java/com/example/meufii/views/activity/UtilFormat.java
@@ -4,7 +4,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 public class UtilFormat {
-    public static String formatDecimal(float number) {
+    public static String formatDecimal(double number) {
         Locale locale = new Locale("pt", "BR");
         NumberFormat numberFormat = NumberFormat.getCurrencyInstance(locale);
         return numberFormat.format(number);
